Move packing to pro-crafting and use a consistent packaging scheme
